Implementd Input Variables
I moved the credentials over to input Variables so they aren't hard
coded in the source.

Also changed the operation of this action to be more of a trigger.
It's passed the booked id and status of the record that changed
and it just needs to authenticate and fire off a status update.